daemon: split aliases support to its own api_*.go files
