+++ daemon/api_test.go
@@ -43,7 +43,6 @@ import (
 	"github.com/snapcore/snapd/asserts/assertstest"
 	"github.com/snapcore/snapd/client"
 	"github.com/snapcore/snapd/dirs"
-	"github.com/snapcore/snapd/osutil"
 	"github.com/snapcore/snapd/overlord/assertstate"
 	"github.com/snapcore/snapd/overlord/assertstate/assertstatetest"
 	"github.com/snapcore/snapd/overlord/auth"
@@ -3606,596 +3605,22 @@ func (s *apiSuite) TestIsTrue(c *check.C) {
 	}
 }
 
-// aliases
+func (s *apiSuite) TestLogsNoServices(c *check.C) {
+	// NOTE this is *apiSuite, not *appSuite, so there are no
+	// installed snaps with services
 
-func (s *apiSuite) TestAliasSuccess(c *check.C) {
-	err := os.MkdirAll(dirs.SnapBinariesDir, 0755)
-	c.Assert(err, check.IsNil)
-	d := s.daemon(c)
-
-	s.MockSnap(c, aliasYaml)
-
-	oldAutoAliases := snapstate.AutoAliases
-	snapstate.AutoAliases = func(*state.State, *snap.Info) (map[string]string, error) {
-		return nil, nil
-	}
-	defer func() { snapstate.AutoAliases = oldAutoAliases }()
-
-	d.overlord.Loop()
-	defer d.overlord.Stop()
-
-	action := &aliasAction{
-		Action: "alias",
-		Snap:   "alias-snap",
-		App:    "app",
-		Alias:  "alias1",
-	}
-	text, err := json.Marshal(action)
-	c.Assert(err, check.IsNil)
-	buf := bytes.NewBuffer(text)
-	req, err := http.NewRequest("POST", "/v2/aliases", buf)
-	c.Assert(err, check.IsNil)
-	rec := httptest.NewRecorder()
-	aliasesCmd.POST(aliasesCmd, req, nil).ServeHTTP(rec, req)
-	c.Assert(rec.Code, check.Equals, 202)
-	var body map[string]interface{}
-	err = json.Unmarshal(rec.Body.Bytes(), &body)
-	c.Check(err, check.IsNil)
-	id := body["change"].(string)
-
-	st := d.overlord.State()
-	st.Lock()
-	chg := st.Change(id)
-	st.Unlock()
-	c.Assert(chg, check.NotNil)
-
-	<-chg.Ready()
-
-	st.Lock()
-	err = chg.Err()
-	st.Unlock()
-	c.Assert(err, check.IsNil)
-
-	// sanity check
-	c.Check(osutil.IsSymlink(filepath.Join(dirs.SnapBinariesDir, "alias1")), check.Equals, true)
-}
-
-func (s *apiSuite) TestAliasChangeConflict(c *check.C) {
-	err := os.MkdirAll(dirs.SnapBinariesDir, 0755)
-	c.Assert(err, check.IsNil)
+	cmd := testutil.MockCommand(c, "systemctl", "").Also("journalctl", "")
+	defer cmd.Restore()
 	s.daemon(c)
+	s.d.overlord.Loop()
+	defer s.d.overlord.Stop()
 
-	s.MockSnap(c, aliasYaml)
-
-	s.SimulateConflict("alias-snap")
-
-	oldAutoAliases := snapstate.AutoAliases
-	snapstate.AutoAliases = func(*state.State, *snap.Info) (map[string]string, error) {
-		return nil, nil
-	}
-	defer func() { snapstate.AutoAliases = oldAutoAliases }()
-
-	action := &aliasAction{
-		Action: "alias",
-		Snap:   "alias-snap",
-		App:    "app",
-		Alias:  "alias1",
-	}
-	text, err := json.Marshal(action)
-	c.Assert(err, check.IsNil)
-	buf := bytes.NewBuffer(text)
-	req, err := http.NewRequest("POST", "/v2/aliases", buf)
-	c.Assert(err, check.IsNil)
-	rec := httptest.NewRecorder()
-	aliasesCmd.POST(aliasesCmd, req, nil).ServeHTTP(rec, req)
-	c.Check(rec.Code, check.Equals, 409)
-
-	var body map[string]interface{}
-	err = json.Unmarshal(rec.Body.Bytes(), &body)
-	c.Check(err, check.IsNil)
-	c.Check(body, check.DeepEquals, map[string]interface{}{
-		"status-code": 409.,
-		"status":      "Conflict",
-		"result": map[string]interface{}{
-			"message": `snap "alias-snap" has "manip" change in progress`,
-			"kind":    "snap-change-conflict",
-			"value": map[string]interface{}{
-				"change-kind": "manip",
-				"snap-name":   "alias-snap",
-			},
-		},
-		"type": "error"})
-}
-
-func (s *apiSuite) TestAliasErrors(c *check.C) {
-	s.daemon(c)
-
-	errScenarios := []struct {
-		mangle func(*aliasAction)
-		err    string
-	}{
-		{func(a *aliasAction) { a.Action = "" }, `unsupported alias action: ""`},
-		{func(a *aliasAction) { a.Action = "what" }, `unsupported alias action: "what"`},
-		{func(a *aliasAction) { a.Snap = "lalala" }, `snap "lalala" is not installed`},
-		{func(a *aliasAction) { a.Alias = ".foo" }, `invalid alias name: ".foo"`},
-		{func(a *aliasAction) { a.Aliases = []string{"baz"} }, `cannot interpret request, snaps can no longer be expected to declare their aliases`},
-	}
-
-	for _, scen := range errScenarios {
-		action := &aliasAction{
-			Action: "alias",
-			Snap:   "alias-snap",
-			App:    "app",
-			Alias:  "alias1",
-		}
-		scen.mangle(action)
-
-		text, err := json.Marshal(action)
-		c.Assert(err, check.IsNil)
-		buf := bytes.NewBuffer(text)
-		req, err := http.NewRequest("POST", "/v2/aliases", buf)
-		c.Assert(err, check.IsNil)
-
-		rsp := changeAliases(aliasesCmd, req, nil).(*resp)
-		c.Check(rsp.Type, check.Equals, ResponseTypeError)
-		c.Check(rsp.Status, check.Equals, 400)
-		c.Check(rsp.Result.(*errorResult).Message, check.Matches, scen.err)
-	}
-}
-
-func (s *apiSuite) TestUnaliasSnapSuccess(c *check.C) {
-	err := os.MkdirAll(dirs.SnapBinariesDir, 0755)
-	c.Assert(err, check.IsNil)
-	d := s.daemon(c)
-
-	s.MockSnap(c, aliasYaml)
-
-	oldAutoAliases := snapstate.AutoAliases
-	snapstate.AutoAliases = func(*state.State, *snap.Info) (map[string]string, error) {
-		return nil, nil
-	}
-	defer func() { snapstate.AutoAliases = oldAutoAliases }()
-
-	d.overlord.Loop()
-	defer d.overlord.Stop()
-
-	action := &aliasAction{
-		Action: "unalias",
-		Snap:   "alias-snap",
-	}
-	text, err := json.Marshal(action)
-	c.Assert(err, check.IsNil)
-	buf := bytes.NewBuffer(text)
-	req, err := http.NewRequest("POST", "/v2/aliases", buf)
-	c.Assert(err, check.IsNil)
-	rec := httptest.NewRecorder()
-	aliasesCmd.POST(aliasesCmd, req, nil).ServeHTTP(rec, req)
-	c.Assert(rec.Code, check.Equals, 202)
-	var body map[string]interface{}
-	err = json.Unmarshal(rec.Body.Bytes(), &body)
-	c.Check(err, check.IsNil)
-	id := body["change"].(string)
-
-	st := d.overlord.State()
-	st.Lock()
-	chg := st.Change(id)
-	c.Check(chg.Summary(), check.Equals, `Disable all aliases for snap "alias-snap"`)
-	st.Unlock()
-	c.Assert(chg, check.NotNil)
-
-	<-chg.Ready()
-
-	st.Lock()
-	defer st.Unlock()
-	err = chg.Err()
+	req, err := http.NewRequest("GET", "/v2/logs", nil)
 	c.Assert(err, check.IsNil)
 
-	// sanity check
-	var snapst snapstate.SnapState
-	err = snapstate.Get(st, "alias-snap", &snapst)
-	c.Assert(err, check.IsNil)
-	c.Check(snapst.AutoAliasesDisabled, check.Equals, true)
-}
-
-func (s *apiSuite) TestUnaliasDWIMSnapSuccess(c *check.C) {
-	err := os.MkdirAll(dirs.SnapBinariesDir, 0755)
-	c.Assert(err, check.IsNil)
-	d := s.daemon(c)
-
-	s.MockSnap(c, aliasYaml)
-
-	oldAutoAliases := snapstate.AutoAliases
-	snapstate.AutoAliases = func(*state.State, *snap.Info) (map[string]string, error) {
-		return nil, nil
-	}
-	defer func() { snapstate.AutoAliases = oldAutoAliases }()
-
-	d.overlord.Loop()
-	defer d.overlord.Stop()
-
-	action := &aliasAction{
-		Action: "unalias",
-		Snap:   "alias-snap",
-		Alias:  "alias-snap",
-	}
-	text, err := json.Marshal(action)
-	c.Assert(err, check.IsNil)
-	buf := bytes.NewBuffer(text)
-	req, err := http.NewRequest("POST", "/v2/aliases", buf)
-	c.Assert(err, check.IsNil)
-	rec := httptest.NewRecorder()
-	aliasesCmd.POST(aliasesCmd, req, nil).ServeHTTP(rec, req)
-	c.Assert(rec.Code, check.Equals, 202)
-	var body map[string]interface{}
-	err = json.Unmarshal(rec.Body.Bytes(), &body)
-	c.Check(err, check.IsNil)
-	id := body["change"].(string)
-
-	st := d.overlord.State()
-	st.Lock()
-	chg := st.Change(id)
-	c.Check(chg.Summary(), check.Equals, `Disable all aliases for snap "alias-snap"`)
-	st.Unlock()
-	c.Assert(chg, check.NotNil)
-
-	<-chg.Ready()
-
-	st.Lock()
-	defer st.Unlock()
-	err = chg.Err()
-	c.Assert(err, check.IsNil)
-
-	// sanity check
-	var snapst snapstate.SnapState
-	err = snapstate.Get(st, "alias-snap", &snapst)
-	c.Assert(err, check.IsNil)
-	c.Check(snapst.AutoAliasesDisabled, check.Equals, true)
-}
-
-func (s *apiSuite) TestUnaliasAliasSuccess(c *check.C) {
-	err := os.MkdirAll(dirs.SnapBinariesDir, 0755)
-	c.Assert(err, check.IsNil)
-	d := s.daemon(c)
-
-	s.MockSnap(c, aliasYaml)
-
-	oldAutoAliases := snapstate.AutoAliases
-	snapstate.AutoAliases = func(*state.State, *snap.Info) (map[string]string, error) {
-		return nil, nil
-	}
-	defer func() { snapstate.AutoAliases = oldAutoAliases }()
-
-	d.overlord.Loop()
-	defer d.overlord.Stop()
-
-	action := &aliasAction{
-		Action: "alias",
-		Snap:   "alias-snap",
-		App:    "app",
-		Alias:  "alias1",
-	}
-	text, err := json.Marshal(action)
-	c.Assert(err, check.IsNil)
-	buf := bytes.NewBuffer(text)
-	req, err := http.NewRequest("POST", "/v2/aliases", buf)
-	c.Assert(err, check.IsNil)
-	rec := httptest.NewRecorder()
-	aliasesCmd.POST(aliasesCmd, req, nil).ServeHTTP(rec, req)
-	c.Assert(rec.Code, check.Equals, 202)
-	var body map[string]interface{}
-	err = json.Unmarshal(rec.Body.Bytes(), &body)
-	c.Check(err, check.IsNil)
-	id := body["change"].(string)
-
-	st := d.overlord.State()
-	st.Lock()
-	chg := st.Change(id)
-	st.Unlock()
-	c.Assert(chg, check.NotNil)
-
-	<-chg.Ready()
-
-	st.Lock()
-	err = chg.Err()
-	st.Unlock()
-	c.Assert(err, check.IsNil)
-
-	// unalias
-	action = &aliasAction{
-		Action: "unalias",
-		Alias:  "alias1",
-	}
-	text, err = json.Marshal(action)
-	c.Assert(err, check.IsNil)
-	buf = bytes.NewBuffer(text)
-	req, err = http.NewRequest("POST", "/v2/aliases", buf)
-	c.Assert(err, check.IsNil)
-	rec = httptest.NewRecorder()
-	aliasesCmd.POST(aliasesCmd, req, nil).ServeHTTP(rec, req)
-	c.Assert(rec.Code, check.Equals, 202)
-	err = json.Unmarshal(rec.Body.Bytes(), &body)
-	c.Check(err, check.IsNil)
-	id = body["change"].(string)
-
-	st.Lock()
-	chg = st.Change(id)
-	c.Check(chg.Summary(), check.Equals, `Remove manual alias "alias1" for snap "alias-snap"`)
-	st.Unlock()
-	c.Assert(chg, check.NotNil)
-
-	<-chg.Ready()
-
-	st.Lock()
-	defer st.Unlock()
-	err = chg.Err()
-	c.Assert(err, check.IsNil)
-
-	// sanity check
-	c.Check(osutil.FileExists(filepath.Join(dirs.SnapBinariesDir, "alias1")), check.Equals, false)
-}
-
-func (s *apiSuite) TestUnaliasDWIMAliasSuccess(c *check.C) {
-	err := os.MkdirAll(dirs.SnapBinariesDir, 0755)
-	c.Assert(err, check.IsNil)
-	d := s.daemon(c)
-
-	s.MockSnap(c, aliasYaml)
-
-	oldAutoAliases := snapstate.AutoAliases
-	snapstate.AutoAliases = func(*state.State, *snap.Info) (map[string]string, error) {
-		return nil, nil
-	}
-	defer func() { snapstate.AutoAliases = oldAutoAliases }()
-
-	d.overlord.Loop()
-	defer d.overlord.Stop()
-
-	action := &aliasAction{
-		Action: "alias",
-		Snap:   "alias-snap",
-		App:    "app",
-		Alias:  "alias1",
-	}
-	text, err := json.Marshal(action)
-	c.Assert(err, check.IsNil)
-	buf := bytes.NewBuffer(text)
-	req, err := http.NewRequest("POST", "/v2/aliases", buf)
-	c.Assert(err, check.IsNil)
-	rec := httptest.NewRecorder()
-	aliasesCmd.POST(aliasesCmd, req, nil).ServeHTTP(rec, req)
-	c.Assert(rec.Code, check.Equals, 202)
-	var body map[string]interface{}
-	err = json.Unmarshal(rec.Body.Bytes(), &body)
-	c.Check(err, check.IsNil)
-	id := body["change"].(string)
-
-	st := d.overlord.State()
-	st.Lock()
-	chg := st.Change(id)
-	st.Unlock()
-	c.Assert(chg, check.NotNil)
-
-	<-chg.Ready()
-
-	st.Lock()
-	err = chg.Err()
-	st.Unlock()
-	c.Assert(err, check.IsNil)
-
-	// DWIM unalias an alias
-	action = &aliasAction{
-		Action: "unalias",
-		Snap:   "alias1",
-		Alias:  "alias1",
-	}
-	text, err = json.Marshal(action)
-	c.Assert(err, check.IsNil)
-	buf = bytes.NewBuffer(text)
-	req, err = http.NewRequest("POST", "/v2/aliases", buf)
-	c.Assert(err, check.IsNil)
-	rec = httptest.NewRecorder()
-	aliasesCmd.POST(aliasesCmd, req, nil).ServeHTTP(rec, req)
-	c.Assert(rec.Code, check.Equals, 202)
-	err = json.Unmarshal(rec.Body.Bytes(), &body)
-	c.Check(err, check.IsNil)
-	id = body["change"].(string)
-
-	st.Lock()
-	chg = st.Change(id)
-	c.Check(chg.Summary(), check.Equals, `Remove manual alias "alias1" for snap "alias-snap"`)
-	st.Unlock()
-	c.Assert(chg, check.NotNil)
-
-	<-chg.Ready()
-
-	st.Lock()
-	defer st.Unlock()
-	err = chg.Err()
-	c.Assert(err, check.IsNil)
-
-	// sanity check
-	c.Check(osutil.FileExists(filepath.Join(dirs.SnapBinariesDir, "alias1")), check.Equals, false)
-}
-
-func (s *apiSuite) TestPreferSuccess(c *check.C) {
-	err := os.MkdirAll(dirs.SnapBinariesDir, 0755)
-	c.Assert(err, check.IsNil)
-	d := s.daemon(c)
-
-	s.MockSnap(c, aliasYaml)
-
-	oldAutoAliases := snapstate.AutoAliases
-	snapstate.AutoAliases = func(*state.State, *snap.Info) (map[string]string, error) {
-		return nil, nil
-	}
-	defer func() { snapstate.AutoAliases = oldAutoAliases }()
-
-	d.overlord.Loop()
-	defer d.overlord.Stop()
-
-	action := &aliasAction{
-		Action: "prefer",
-		Snap:   "alias-snap",
-	}
-	text, err := json.Marshal(action)
-	c.Assert(err, check.IsNil)
-	buf := bytes.NewBuffer(text)
-	req, err := http.NewRequest("POST", "/v2/aliases", buf)
-	c.Assert(err, check.IsNil)
-	rec := httptest.NewRecorder()
-	aliasesCmd.POST(aliasesCmd, req, nil).ServeHTTP(rec, req)
-	c.Assert(rec.Code, check.Equals, 202)
-	var body map[string]interface{}
-	err = json.Unmarshal(rec.Body.Bytes(), &body)
-	c.Check(err, check.IsNil)
-	id := body["change"].(string)
-
-	st := d.overlord.State()
-	st.Lock()
-	chg := st.Change(id)
-	c.Check(chg.Summary(), check.Equals, `Prefer aliases of snap "alias-snap"`)
-	st.Unlock()
-	c.Assert(chg, check.NotNil)
-
-	<-chg.Ready()
-
-	st.Lock()
-	defer st.Unlock()
-	err = chg.Err()
-	c.Assert(err, check.IsNil)
-
-	// sanity check
-	var snapst snapstate.SnapState
-	err = snapstate.Get(st, "alias-snap", &snapst)
-	c.Assert(err, check.IsNil)
-	c.Check(snapst.AutoAliasesDisabled, check.Equals, false)
-}
-
-func (s *apiSuite) TestAliases(c *check.C) {
-	d := s.daemon(c)
-
-	st := d.overlord.State()
-	st.Lock()
-	snapstate.Set(st, "alias-snap1", &snapstate.SnapState{
-		Sequence: []*snap.SideInfo{
-			{RealName: "alias-snap1", Revision: snap.R(11)},
-		},
-		Current: snap.R(11),
-		Active:  true,
-		Aliases: map[string]*snapstate.AliasTarget{
-			"alias1": {Manual: "cmd1x", Auto: "cmd1"},
-			"alias2": {Auto: "cmd2"},
-		},
-	})
-	snapstate.Set(st, "alias-snap2", &snapstate.SnapState{
-		Sequence: []*snap.SideInfo{
-			{RealName: "alias-snap2", Revision: snap.R(12)},
-		},
-		Current:             snap.R(12),
-		Active:              true,
-		AutoAliasesDisabled: true,
-		Aliases: map[string]*snapstate.AliasTarget{
-			"alias2": {Auto: "cmd2"},
-			"alias3": {Manual: "cmd3"},
-			"alias4": {Manual: "cmd4x", Auto: "cmd4"},
-		},
-	})
-	st.Unlock()
-
-	req, err := http.NewRequest("GET", "/v2/aliases", nil)
-	c.Assert(err, check.IsNil)
-
-	rsp := getAliases(aliasesCmd, req, nil).(*resp)
-	c.Check(rsp.Type, check.Equals, ResponseTypeSync)
-	c.Check(rsp.Status, check.Equals, 200)
-	c.Check(rsp.Result, check.DeepEquals, map[string]map[string]aliasStatus{
-		"alias-snap1": {
-			"alias1": {
-				Command: "alias-snap1.cmd1x",
-				Status:  "manual",
-				Manual:  "cmd1x",
-				Auto:    "cmd1",
-			},
-			"alias2": {
-				Command: "alias-snap1.cmd2",
-				Status:  "auto",
-				Auto:    "cmd2",
-			},
-		},
-		"alias-snap2": {
-			"alias2": {
-				Command: "alias-snap2.cmd2",
-				Status:  "disabled",
-				Auto:    "cmd2",
-			},
-			"alias3": {
-				Command: "alias-snap2.cmd3",
-				Status:  "manual",
-				Manual:  "cmd3",
-			},
-			"alias4": {
-				Command: "alias-snap2.cmd4x",
-				Status:  "manual",
-				Manual:  "cmd4x",
-				Auto:    "cmd4",
-			},
-		},
-	})
-
-}
-
-func (s *apiSuite) TestInstallUnaliased(c *check.C) {
-	var calledFlags snapstate.Flags
-
-	snapstateInstall = func(ctx context.Context, s *state.State, name string, opts *snapstate.RevisionOptions, userID int, flags snapstate.Flags) (*state.TaskSet, error) {
-		calledFlags = flags
-
-		t := s.NewTask("fake-install-snap", "Doing a fake install")
-		return state.NewTaskSet(t), nil
-	}
-
-	d := s.daemon(c)
-	inst := &snapInstruction{
-		Action: "install",
-		// Install the snap without enabled automatic aliases
-		Unaliased: true,
-		Snaps:     []string{"fake"},
-	}
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-	_, _, err := inst.dispatch()(inst, st)
-	c.Check(err, check.IsNil)
-
-	c.Check(calledFlags.Unaliased, check.Equals, true)
-}
-
-func (s *apiSuite) TestInstallIgnoreRunning(c *check.C) {
-	var calledFlags snapstate.Flags
-
-	snapstateInstall = func(ctx context.Context, s *state.State, name string, opts *snapstate.RevisionOptions, userID int, flags snapstate.Flags) (*state.TaskSet, error) {
-		calledFlags = flags
-
-		t := s.NewTask("fake-install-snap", "Doing a fake install")
-		return state.NewTaskSet(t), nil
-	}
-
-	d := s.daemon(c)
-	inst := &snapInstruction{
-		Action: "install",
-		// Install the snap without enabled automatic aliases
-		IgnoreRunning: true,
-		Snaps:         []string{"fake"},
-	}
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-	_, _, err := inst.dispatch()(inst, st)
-	c.Check(err, check.IsNil)
-
-	c.Check(calledFlags.IgnoreRunning, check.Equals, true)
+	rsp := getLogs(logsCmd, req, nil).(*resp)
+	c.Assert(rsp.Status, check.Equals, 404)
+	c.Assert(rsp.Type, check.Equals, ResponseTypeError)
 }
 
 func (s *apiSuite) TestInstallPathUnaliased(c *check.C) {
@@ -4218,24 +3643,6 @@ func (s *apiSuite) TestInstallPathUnaliased(c *check.C) {
 	flags := snapstate.Flags{Unaliased: true, RemoveSnapPath: true, DevMode: true}
 	chgSummary := s.sideloadCheck(c, body, head, "local", flags)
 	c.Check(chgSummary, check.Equals, `Install "local" snap from file "x"`)
-}
-
-func (s *apiSuite) TestLogsNoServices(c *check.C) {
-	// NOTE this is *apiSuite, not *appSuite, so there are no
-	// installed snaps with services
-
-	cmd := testutil.MockCommand(c, "systemctl", "").Also("journalctl", "")
-	defer cmd.Restore()
-	s.daemon(c)
-	s.d.overlord.Loop()
-	defer s.d.overlord.Stop()
-
-	req, err := http.NewRequest("GET", "/v2/logs", nil)
-	c.Assert(err, check.IsNil)
-
-	rsp := getLogs(logsCmd, req, nil).(*resp)
-	c.Assert(rsp.Status, check.Equals, 404)
-	c.Assert(rsp.Type, check.Equals, ResponseTypeError)
 }
 
 type fakeNetError struct {
